Materials/Containerized Flask Prod/requirements.txt
flask==2.1.2
psycopg2-binary==2.9.3
psycopg2==2.9.3
psycopg~=3.0.8
psycopg-c~=3.0.8
redis==4.2.2
yoyo-migrations==7.3.2
asgiref==3.5.2
python-dotenv==0.20.0
gunicorn==20.1.0
gevent==21.12.0